# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ComicClan",
        "description": "An complete API made to serve new upcoming ComicClan social network",
        "contact": {
            "name": "Jean Melo",
            "email": "jf.melo6@gmail.com"
        },
        "version": "0.0.1-oas3"
    },
    "servers": [
        {
            "url": "https://virtserver.swaggerhub.com/ghanizadev/comic-clan/1.0.0"
        }
    ],
    "security": [
        {
            "OAuth2 Authentication": []
        }
    ],
...
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all posts",
                "description": "It includes all comments and media attached to its posts. It can be used as \"feed\" as well.",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current post's page",
                        "schema":{
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Amount of posts per page",
                        "schema":{
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Amount of posts skipped in first page",
                        "schema":{
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort posts by date or description. It can be ascending e.g:\"sort=createdAt\" or descending e.g: \"sort=-createdAt\"",
                        "schema":{
                            "type": "string",
                            "enum": [
                                "createdAt",
                                "updatedAt",
                                "description"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Operation successfully completed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedPosts"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - It is only possible to change your own data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type - Request is contains an unsupported media type. It must be \"application/json\".",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2 Authentication": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a post",
                "description": "Create a single post.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAlterPost"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created - Post was successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedPosts"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - It is only possible to change your own data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type - Request is contains an unsupported media type. It must be \"application/json\".",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2 Authentication": []
                    }
                ]
            }
        },
        "/posts/{postId}":{
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Alter a post",
                "description": "Change a post data, including media assiciation.",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Post unique ID",
                        "schema": {
                            "type": "string",
                            "format": "hex"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/CreateAlterPost"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created - Post was successfully changed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedPosts"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - \"posts\" is not included in your scope.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "description": "Completly remove a post from database. This action is not reversible.",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Post unique ID",
                        "schema":{
                            "type": "string",
                            "format": "hex"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Post was successfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - It is only possible to delete your own posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{postId}/comment":{
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Comment to a post",
                "description": "Submits a new comment to a post.",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Post unique ID",
                        "schema": {
                            "type": "string",
                            "format": "hex"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/CreateAlterComment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created - Post was successfully changed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type - Request is contains an unsupported media type. It must be \"application/json\".",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2 Authentication" : ["feed", "comment"]
                    }
                ]
            }
        },
        "/posts/{postId}/subscribe":{
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Subscribe to a post",
                "description": "Subscribes yourself to receive an email at any post update.",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Post unique ID",
                        "schema": {
                            "type": "string",
                            "format": "hex"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - Subscribed successfully to this post."
                    },
                    "401": {
                        "description": "Unauthorized - \"posts\" is not included in your scope.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedPosts": {
                "type": "object",
                "properties": {
                    "docs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    },
                    "totalDocs": {
                        "type": "number",
                        "example": 20
                    },
                    "limit": {
                        "type": "number",
                        "default": 10
                    },
                    "totalPages": {
                        "type": "number",
                        "example": 2
                    },
                    "page": {
                        "type": "number",
                        "example": 1
                    },
                    "pagingCounter": {
                        "type": "number",
                        "example": 1
                    },
                    "hasPrevPage": {
                        "type": "boolean",
                        "example": false
                    },
                    "hasNextPage": {
                        "type": "boolean",
                        "example": true
                    },
                    "prevPage": {
                        "type": "number"
                    },
                    "nextPage": {
                        "type": "number",
                        "example": 2
                    }
                },
                "xml": {
                    "name": "PaggingPost"
                }
            },
...
            "Post": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "format": "hex"
                    },
                    "description": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "xml": {
                    "name": "Post"
                }
            },
            "CreateAlterPost": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "xml": {
                    "name": "Post"
                }
            },
            "CreateAlterComment": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "xml": {
                    "name": "CreateAlterComment"
                }
            },
            "Comment": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "format": "hex"
                    },
                    "body": {
                        "type": "string"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Reply"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "xml": {
                    "name": "Post"
                }
            },
            "Reply": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "format": "hex"
                    },
                    "body": {
                        "type": "string"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "xml": {
                    "name": "Post"
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "error_description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "number"
                    }
                },
                "xml": {
                    "name": "Error"
                }
            },
...
        },
...
        "securitySchemes": {
            "Basic Authentication": {
                "type": "http",
                "scheme": "basic"
            },
            "OAuth2 Authentication": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "https://localhost/oauth/token",
                        "refreshUrl": "https://localhost/oauth/token",
                        "scopes": {
                            "feed": "allows reading feed resources",
                            "post": "allows posting to user's feed",
                            "profile": "allows reading, editing and deleting user's profile",
                            "comment": "allows commenting in post resources"
                        }
                    }
                }
            }
        }
    }
}